Trap code cleanup.
