+++ src/cave.h
@@ -76,7 +76,6 @@ extern struct cave *cave_new(void);
 extern void cave_free(struct cave *c);
 
 extern void cave_set_feat(struct cave *c, int y, int x, int feat);
-extern void cave_set_trap(struct cave *c, int y, int x, struct trap *t_ptr);
 extern void cave_note_spot(struct cave *c, int y, int x);
 extern void cave_light_spot(struct cave *c, int y, int x);
 extern void cave_update_flow(struct cave *c);
