+++ src/externs.h
@@ -10,6 +10,7 @@
 #include "z-file.h"
 #include "z-msg.h"
 #include "spells.h"
+#include "trap.h"
 
 /* This file was automatically generated. It is now obsolete (it was never a
  * good idea to begin with; you should include only what you use instead of
