+++ src/types.h
@@ -101,44 +101,6 @@ typedef struct feature
 	wchar_t x_char[3];   /**< Desired feature character (set by user/pref file) */
 } feature_type;
 
-/**
- * Information about trap kinds.
- */
-typedef struct trap_kind
-{
-	char *name;
-	s16b idx;
-
-	struct trap_kind *next;
-
-	int hidden;		/* How hidden is the trap? */
-
-	int min_level;
-	int max_level;
-	
-	u32b effect;   /**< Effect on entry to grid */
-
-	byte d_attr;   /**< Default feature attribute */
-	wchar_t d_char;   /**< Default feature character */
-
-	byte x_attr[3];   /**< Desired feature attribute (set by user/pref file) */
-	wchar_t x_char[3];   /**< Desired feature character (set by user/pref file) */
-} trap_kind;
-
-/**
- * Information about one particular trap.
- */
-typedef struct trap
-{
-	struct trap_kind *kind;
-	
-	u16b hidden;
-	
-	byte x;
-	byte y;
-
-} trap_type;
-
 /*
  * Information about "vault generation"
  */
